Moving hangout creation logic out from Meteor method so it can be reused from bot
